src/requirements.txt
environs==7.2.0
sqlalchemy==1.3.13
cx-Oracle==7.3.0
factory-boy==2.12.0
pytest==5.3.5